requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==5.0.6
django-allauth==0.63.3
django-crispy-forms==2.2
djangorestframework==3.15.2
idna==3.7
oauthlib==3.2.2
pillow==10.4.0
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.0
stripe==10.2.0
typing_extensions==4.12.2
urllib3==2.2.2
